+++ partials/mobile/mobile-dropdown.php
@@ -75,7 +75,7 @@ $close_text = $close_text ? $close_text : esc_html__( 'Close', 'oceanwp' );
 
 <div id="mobile-dropdown" class="clr" <?php echo $dropdown_menu_attrs; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>>
 
-	<nav class="<?php echo esc_attr( $classes ); ?>"<?php oceanwp_schema_markup( 'site_navigation' ); ?>>
+	<nav class="<?php echo esc_attr( $classes ); ?>"<?php oceanwp_schema_attr( 'site_navigation' ); ?>>
 
 		<?php
 		// If has mobile menu.