+++ partials/single/content.php
@@ -14,7 +14,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 <?php do_action( 'ocean_before_single_post_content' ); ?>
 
-<div class="entry-content clr"<?php oceanwp_schema_markup( 'entry_content' ); ?>>
+<div class="entry-content clr"<?php oceanwp_schema_attr( 'entry_content' ); ?>>
 	<?php
 	the_content();