+++ templates/landing.php
@@ -17,7 +17,7 @@
 	</head>
 
 	<!-- Begin Body -->
-	<body <?php body_class(); ?> <?php oceanwp_schema_markup( 'html' ); ?>>
+	<body <?php body_class(); ?> <?php oceanwp_schema_attr( 'html' ); ?>>
 
 		<?php wp_body_open(); ?>
 
@@ -33,7 +33,7 @@
 
 				<?php do_action( 'ocean_before_main' ); ?>
 
-				<main id="main" class="site-main clr"<?php oceanwp_schema_markup( 'main' ); ?> role="main">
+				<main id="main" class="site-main clr"<?php oceanwp_schema_attr( 'main' ); ?> role="main">
 
 					<?php do_action( 'ocean_before_content_wrap' ); ?>